Add program course lists and use float credits
Introduce support for program-level course lists and standardize credits as floats. Changes include: models — Course.credits -> Float, rename Program.total_credits -> program_credits and add ProgramCourseLists model + relationship; schemas — add ProgramCourseListOut and update ProgramStructureOut to use program_credits and include course_lists; queries — eager-load Program.course_lists and related Course rows; routers — assemble course_lists in program structure response and propogate program_credits rename; scrapers — course_scraper: add build_session(), parse credits as float and use Float32 in DataFrame, adjust scrape flow and CSV export; program_scraper: add _extract_course_lists(), switch to program_credits, include course_lists in extracted data and change output CSV name; services — program_builder: add build_list() to convert scraped lists into ORM rows and wire into program building, use program_credits for calculations; writers — clear and write program_course_lists in DB operations and handle float credits; update_database — add optional program link scraping step and adjust run args. Also fix parse_requirements return to return prerequisites when no headers. Overall this update adds course-list support across scraper, builder, schema, query and persistence layers and aligns credit handling to floats.

+++ backend/app/schemas/course.py
@@ -6,7 +6,7 @@ class Course(BaseModel):
     course_id: int
     course_code: str
     title: Optional[str]
-    credits: Optional[int]
+    credits: Optional[float]
     description: Optional[str]
     prerequisite_str: Optional[str] = None
     # How this course appears in a program:
